fix for #2743 : When a Task is hanging, it can prevent other tasks from running on the same JVM - cache cleaning is now performed using a trylock call. Without this change the cache cleaning thread could block the execution of tasks on the node while its trying to acquire the node. As the cache cleaning is supposed to be performed only when a node is idle, trylock fits well. - added a regression test which reproduces the problem.

+++ rm/rm-node/src/main/java/org/ow2/proactive/resourcemanager/nodesource/dataspace/DataSpaceNodeConfigurationAgent.java
@@ -302,18 +302,19 @@ public class DataSpaceNodeConfigurationAgent implements Serializable {
                 long invalidationPeriod = getCacheInvalidationPeriod();
                 long currentTime = System.currentTimeMillis();
                 // lock the timer in write mode, this will prevent any Task to start during the cleaning process
-                cacheCleaningRWLock.writeLock().lockInterruptibly();
-                try {
-                    FileObject rootFO = fileSystemManager.resolveFile(rootCacheUri);
-                    FileObject[] files = rootFO.findFiles(Selectors.EXCLUDE_SELF);
-                    for (FileObject file : files) {
-                        if (currentTime - file.getContent().getLastModifiedTime() > invalidationPeriod) {
-                            logger.info("[Cache Space cleaner] deleting " + file);
-                            file.delete();
+                if (cacheCleaningRWLock.writeLock().tryLock()) {
+                    try {
+                        FileObject rootFO = fileSystemManager.resolveFile(rootCacheUri);
+                        FileObject[] files = rootFO.findFiles(Selectors.EXCLUDE_SELF);
+                        for (FileObject file : files) {
+                            if (currentTime - file.getContent().getLastModifiedTime() > invalidationPeriod) {
+                                logger.info("[Cache Space cleaner] deleting " + file);
+                                file.delete();
+                            }
                         }
+                    } finally {
+                        cacheCleaningRWLock.writeLock().unlock();
                     }
-                } finally {
-                    cacheCleaningRWLock.writeLock().unlock();
                 }
             } catch (Exception e) {
                 logger.error("Error when cleaning files in cache", e);
